+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/Basic/Left to Right/VisualToastNotificationBasicForm.resx
