+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/Basic/Right to Left/VisualToastNotificationBasicRtlAwareForm.cs
@@ -1,0 +1,36 @@
+﻿#region BSD License
+/*
+ *
+ *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
+ *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2024 - 2024. All rights reserved.
+ *
+ */
+#endregion
+
+using ContentAlignment = System.Drawing.ContentAlignment;
+using Timer = System.Windows.Forms.Timer;
+
+namespace Krypton.Toolkit
+{
+    internal partial class VisualToastNotificationBasicRtlAwareForm : KryptonForm
+    {
+        #region Instance Fields
+
+        private int _time;
+
+        private Timer _timer;
+
+        private SoundPlayer? _soundPlayer;
+
+        private PaletteBase _palette;
+
+        private readonly KryptonBasicToastNotificationData _basicToastNotificationData;
+
+        #endregion
+
+        public VisualToastNotificationBasicRtlAwareForm()
+        {
+            InitializeComponent();
+        }
+    }
+}
