+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/Basic/Right to Left/VisualToastNotificationBasicRtlAwareForm.resx
