+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/Notification Toast/User Input/Left to Right/VisualToastNotificationUserInputForm.resx
